First compiling implementation of getifaddrs()
